+++ debug/xtream-codec-core-debug/src/main/java/io/github/hylexus/xtream/debug/codec/core/demo02/RustStyleDebugEntity02Flatten.java
@@ -40,7 +40,7 @@ public class RustStyleDebugEntity02Flatten {
 
     // byte[4]     协议版本号
     @Preset.RustStyle.u8
-    private byte protocolVersion;
+    private short protocolVersion;
 
     // byte[5-15)    终端手机号或设备ID bcd[10]
     @Preset.RustStyle.str(charset = XtreamConstants.CHARSET_NAME_BCD_8421, length = 10)
